refactor: move NewVehicleForm to new dir
