+++ src/pages-components/Dashboard/FleetPage/NewVehicleForm/index.ts
@@ -1,0 +1,1 @@
+export { NewVehicleForm } from './NewVehicleForm';
