+++ src/pages-components/Dashboard/FleetPage/NewVehicleForm/schema.ts
